+++ gopca/plotting/plot_signature_matrix.py
@@ -113,7 +113,7 @@ def main(args=None):
             verbose = verbose)
 
     # read GO-PCA output
-    G = util.read_gopca_output(gopca_file)
+    G = util.read_gopca_result(gopca_file)
 
     signatures = G.signatures
     S = G.S