+++ gopca/plotting/plot_term_by_pc_matrix.py
@@ -137,7 +137,7 @@ def main(args=None):
     #logger = misc.configure_logger('')
 
     # read GO-PCA output
-    G = util.read_gopca_output(gopca_file)
+    G = util.read_gopca_result(gopca_file)
 
     genes = G.genes
     signatures = G.signatures