+++ gopca/run.py
@@ -21,34 +21,36 @@ import logging
 from copy import deepcopy
 import cPickle as pickle
 
-from gopca import GOPCAConfig, GOPCAOutput
+from gopca import GOPCAConfig, GOPCAResult
 
 logger = logging.getLogger(__name__)
 
 class GOPCARun(object):
-    """A GO-PCA run, consisting of configuration, run, and output data.
+    """A GO-PCA run, consisting of configuration, run, and result data.
 
     Parameters
     ----------
+    version: str
+        The GO-PCA version.
     user_config: `gopca.GOPCAConfig`
         The configuration data provided by the user.
-    genes: tuple or list
-        The list of genes (gene symbols) in the analysis.
-    samples: tuple or list
-        The list of samples (sample names) in the analysis.
+    timestamp: str
+        The timestamp.
+    result: `GOPCAResult`
+        The result (signatures).
     """
 
-    def __init__(self, version, user_config, timestamp, output):
+    def __init__(self, version, user_config, timestamp, result):
         # checks
         assert isinstance(version, (str, unicode))
         assert isinstance(user_config, GOPCAConfig)
         assert isinstance(timestamp, (str, unicode))
-        assert isinstance(output, GOPCAOutput)
+        assert isinstance(result, GOPCAResult)
         # initialization
         self.version = version
         self.user_config = deepcopy(user_config)
         self.timestamp = timestamp
-        self.output = output
+        self.result = result
 
     ### magic functions
     def __repr__(self):
@@ -57,7 +59,7 @@ class GOPCARun(object):
 
     def __str__(self):
         return '<GOPCARun (version %s; %d signatures; %s)>' \
-                %(self.output.q, self.timestamp)
+                %(self.result.q, self.timestamp)
 
     def __eq__(self,other):
         if type(self) is not type(other):
@@ -68,9 +70,9 @@ class GOPCARun(object):
     def __hash__(self):
         data = []
         data.append(hash(self.user_config))
-        data.append(version)
+        data.append(self.version)
         data.append(self.timestamp)
-        data.append(hash(self.output))
+        data.append(hash(self.result))
         return hash(tuple(data))
     ### end magic functions