+++ gopca/scripts/extract_signature_matrix.py
@@ -66,7 +66,7 @@ def main(args=None):
     logger = misc.get_logger(log_file = log_file, quiet = quiet,
             verbose = verbose)
 
-    G = util.read_gopca_output(gopca_file)
+    G = util.read_gopca_result(gopca_file)
     
     signatures = G.signatures
     labels = [sig.get_label(include_id=False) for sig in signatures]