+++ gopca/scripts/extract_signatures.py
@@ -62,7 +62,7 @@ def main(args=None):
 
     assert os.path.isfile(gopca_file)
 
-    G = util.read_gopca_output(gopca_file)
+    G = util.read_gopca_result(gopca_file)
     signatures = G.signatures
 
     # sort signatures first by PC, then by fold enrichment